Added documentation! Mostly ready to go
