+++ chiselFrontend/src/main/scala/chisel3/aop/Aspect.scala
@@ -1,52 +1,34 @@
 package chisel3.aop
 
-import chisel3.core.{BaseModule, Data, Module, ModuleAspect, MultiIOModule, RawModule, RunFirrtlTransform, withClockAndReset}
-import chisel3.internal.{Builder, HasId}
-import chisel3.internal.firrtl.DefModule
-import firrtl.{AnnotationSeq, RenameMap, Transform}
-import firrtl.annotations._
-import firrtl.ir.EmptyStmt
+import chisel3.core.RawModule
+import firrtl.AnnotationSeq
 
 import scala.reflect.runtime.universe.TypeTag
 
-object Aspect {
-  def getFirrtl(chiselIR: chisel3.internal.firrtl.Circuit): Seq[firrtl.ir.DefModule] = {
-    chisel3.internal.firrtl.Converter.convert(chiselIR).modules
-  }
-}
-
-/**
-  *
+/** Represents an aspect of a Chisel module, by specifying
+  *  (1) how to select a Chisel instance from the design
+  *  (2) what behavior should be done to selected instance, via the FIRRTL Annotation Mechanism
   * @param selectRoot Given top-level module, pick the module to apply the aspect (root module)
-  * @param tag Needed for reasons
+  * @param tag Needed to prevent type-erasure of the top-level module type
   * @tparam DUT Type of top-level module
   * @tparam M Type of root module (join point)
   */
 abstract class Aspect[DUT <: RawModule, M <: RawModule](selectRoot: DUT => M)(implicit tag: TypeTag[DUT]) {
+  /** Convert this Aspect to a seq of FIRRTL annotation
+    * @param dut
+    * @return
+    */
   def toAnnotation(dut: DUT): AnnotationSeq
 }
 
+/** Holds utility functions for Aspect stuff */
+object Aspect {
 
-abstract class AnnotatingAspect[DUT <: RawModule, M <: RawModule](selectRoot: DUT => M, selectSignals: M => Seq[Data])(implicit tag: TypeTag[DUT]) extends Aspect[DUT, M](selectRoot) {
-  def toTargets(dut: DUT): Seq[ReferenceTarget] = selectSignals(selectRoot(dut)).map(_.toTarget)
-  def toAnnotation(dut: DUT): AnnotationSeq
-}
-
-
-abstract class Concern[T <: RawModule, R <: Aspect[T, _]](implicit tag: TypeTag[T]) extends Annotation with RunFirrtlTransform {
-
-  def aspects: Seq[R]
-
-  def transformClass: Class[_ <: Transform]
-
-  def toAnnotation(dut: T): AnnotationSeq = aspects.flatMap(_.toAnnotation(dut))
-
-  def resolveAspects(dut: RawModule): AnnotationSeq = {
-    aspects.flatMap(_.toAnnotation(dut.asInstanceOf[T]))
+  /** Converts elaborated Chisel components to FIRRTL modules
+    * @param chiselIR
+    * @return
+    */
+  def getFirrtl(chiselIR: chisel3.internal.firrtl.Circuit): Seq[firrtl.ir.DefModule] = {
+    chisel3.internal.firrtl.Converter.convert(chiselIR).modules
   }
-
-  override def update(renames: RenameMap): Seq[Annotation] = Seq(this)
-  override def toFirrtl: Annotation = this
 }
-
-
