+++ chiselFrontend/src/main/scala/chisel3/aop/DesignAnnotation.scala
@@ -1,10 +1,16 @@
 package chisel3.aop
 
 import chisel3.core.{RawModule, RunFirrtlTransform}
+import firrtl.annotations.NoTargetAnnotation
+import firrtl.Transform
 
-case class DesignAnnotation[T <: RawModule](design: T) extends RunFirrtlTransform with _root_.firrtl.annotations.NoTargetAnnotation {
-  override def transformClass: Class[_ <: _root_.firrtl.Transform] = classOf[_root_.firrtl.Transform]
-  override def toFirrtl = this
+/** Contains the top-level elaborated Chisel design.
+  *
+  * By default is created during Chisel elaboration and passed to the FIRRTL compiler.
+  * @param design top-level Chisel design
+  * @tparam DUT Type of the top-level Chisel design
+  */
+case class DesignAnnotation[DUT <: RawModule](design: DUT) extends RunFirrtlTransform with NoTargetAnnotation {
+  override def transformClass: Class[_ <: Transform] = classOf[Transform]
+  override def toFirrtl: DesignAnnotation[DUT] = this
 }
-
-
